[test] Improve cov.: Changed passed id_card2 value for account details
In the account controller, in order to test if the values we pass are valid files, we check whether they have a filename attribute.
If not, then they are replaced by a False value.

So, we use a deliberatly invalid value for id_card2, in order to trigger this detection and increase coverage.