Extracted support project declaration from commown_self_troubleshooting
We inherited commown_self_troubleshooting purely for the declaration of the support.
However, it's the only element we need, and  we don't require other heavy dependencies
that commown_self_troubleshooting itself inherits,especially sale and website from product_rental.

As such, we decided to extract the project and to insert it here.