Modified action check to verify if a model is abstract
In Odoo 16.0 was added a secondary mixin class in the utm module, utm.source.mixin,
which allows to add only a source to a record without a campaign or medium.

To account for that and potential future mixin classes being added,
we now check if the provided model is abstract to avoid trying to access its attributes.