[test] Replaced CRM demo record call by a dummy model record
Beforehand, the module necessitated a dependency to the CRM module, because we used a demo crm.lead in the test,
in order to assert the success of the sources merge.

Since this is the only use, it's better to declare a dummy model which inherits utm.mixin, and to run tests on records from that model.