[test] Adapted test to use Werkzeug client and methods
Beforehand, since we were using manually the HttpCase request methods, and encoding the parameters in the request url,
it meant the get_form/get_page methods weren't used, and it reduced lisibility.

To standardize the test with other similar tests and to use these previously unused methods, we now create a Werkzeug client and perform HTTP requests through it.