[test] Added inject_payment_data call if account_invoice_merge_auto_pay is installed
The account_invoice_merge_auto_pay module adds a mechanism to automatically merge and pay invoices,
which is crashes if the auto_merge attribute is set to True (set in the contract_auto_merge_invoice module)
and the payment mode of the invoice is not set.
However, depending on the used modules, there might be no payment mode associated with the environnement company.
As such, when this module is installed, we call the inject_payment_data method to automatically create a payment mode for the enrivonnement company, and the company partner we create.