[coverage] Changed signup_retrieve_info override to overload
Beforehand, we entirely rewrote the signup_restrieve_info method in order to replace the instructions
related to name, by instructions for firstname/lastname.
However, this does mean covering lines that are unrelated to the feature we're trying to add.

As such, we simply call the parent class method, then add firstname/lastname in the response.