[test] Added date_start overwrite to match recurring_next_date
When computing next_period_date_end, it computes next_period_date_start,
which is based on the contract line date_start field if last_date_invoiced is not set.
As such, it calculates not from the wanted recurring_next_date, but from date_start.

To avoid that, we simply set the contract date_start to the recurring_next_date value, to fill the data
to match the wanted period.