[test] Added Slimpay provider assignment to payment method when account_payment is installed
When the account_payment module is installed (alongside the sale module), a new field
is added to account.payment.method, payment_provider_id.
Then, when creating transactions, any payment methods which don't match the provider assigned in the transaction
are filtered out (see odoo/addons/account_payment/models/payment_transaction.py:L139).

As such, when the module is installed, we assign the Slimpay provider to the journal payment methods
to be able to pass the tests.