Fixed mail.channel.member search
When an onchange method is called, a cloned NewId record of the user is passed.
However, in Odoo 16.0, the related records attributes, including partner_id, no longer point to the actual records,
but are likewise cloned NewId type records from the original records, meaning it's not possible to directly use self for either the user or its partner value to perform the operations.

As such, since we can consider the users to be identical, we simply pass the partner_id of the origin user record.