[test] Removed assertTrue send_email_on_delivery from tests
Through the commown_shipping_crm_contract_delivery_mail module, the send_email_on_delivery field value on crm.lead
can be different between the default value and the value on the created lead.

As such, to avoid a crash in a multi module environnement, it's better to assign the value we wish right away.

We still need to test the default value of send_email_on_delivery, however it would be best to test it in its own test method.