Removed redirection handling when an exception is encountered
The replacement for _handle_exception is _handle_error.
In _handle_exception (v12), a redirection detection was only added in the case of a 404 Error, but not a 403 Unauthorized error, explaining this overload.

However, in v16, _handle_error tries to find a redirection for both 404 and 403 HTTP errors.
As such, we're no longer required to maintain this code.