Do not use wizard 'usage' related field in methods called from onchange
In 12.0 it simply does not work and return False for permission-related
(bad) reasons it seems. Note that in onchange, the entity is a NewId,
which might be the cause of the bug.

Also note that in 12.0, we cannot explicitely set permissions on transient
models (like the concerned crm.lead.to.customer.wizard), which changed in
16.0 so this should work in 16.0.