Add an action to primary service storable config to copy the image
The button simply calls a method that copies the service image
into the storable template image.

The template image is copied to the variant's by the corresponding
core module during this operation.

A user error is raised if the button is pressed while the service
product still has no image.