Do not reveal contract name to customers when it is not theirs
When the contract associated with a device is not related to a partner's
company, it is no longer displayed.
To implement this bug fix, the contract_name field had to be turned into
a (still non-stored) compute instead of a simple related, so that we can
test the device contract's partner to decide if the contract name should
be exposed or not.