Modified the default selected payment option in the payment screen
We wish to hide the payment options (both providers and tokens) to simplify the payment process, and we reuse the same token for different orders, unless the customer hasn't got any current token or they inserted the SEPA mandate item in their cart.
However, whenever there's several payment providers or tokens, no default payment option is set.

As such, we hardcode the Slimpay provider as the default payment option, which is overriden within the controller if the sepa item is present.