Replace device assignment 'product_id' by 'product_name' to make it customer searchable
The 'product.product' model is indeed not readable by customers so
its name is not searchable. We do not need more than its name so we
store the product name directly, which makes it searchable.

Note the product name is computed when the assignment is created and
does never change, so we should never have a permission read error
when it is computed, hence the absence of sudo_compute=True.