[test] Added a payment method linked to Slimpay
The account.payment record requires a account.payment.method.line record related to a payment method.
In the _create_payment method, the transaction looks for payment methods which match its provider to use as default, and if none are found, the field remains empty.

As such, we create this demo payment method to allow a valid payment creation.