[test] Changed company (or its country) assignment in tests affecting the company
With the multi-company structure of Odoo 16.0, it's required to alter the environnement company or to regenerate its data
in order to pass tests because of account demo data generation.

However, with the assignement of a default country with the commown module based on the environnement company,
newly created partners can have base.fr (France) as a default company - this is an issue because the _slimpay_check_zip constraint method,
triggered upon partner creation, checks if the partner's country is France, and cause an issue if the zip code is unset.

In order to avoid this, we either removed the country from the company, or regenerated data for the default company, in order to avoid this issue.