Fix shop page price for rental products
Before the patch, the rental price inserted in the html page was
not accounting for the eventual pricelist reduction.
This was not really a problem as the javascript code fixes it, so
the user was not impacted.
It is however interesting to fix it to ease unit-testing. It also
simplifies the code a bit.