[test] Removed now useless test for writing on companies as a cust. admin
This test was written during the 16.0 migration of the customer_manager_base module.
However, a new PR in the 12.0 version of this module, which was ported after the migration of this module,
added a clause to forbid customer admins to write on company res.partner records.
(See 'res_partner_customer_user_rule' ir.rule record in the module data)

This behavior is tested in the method 'test_rule_write_not_granted_on_company_partners' added alongside this change.
As such, this test is now useless.