Do not return instance with sudo env to avoid further privilege escalation
When creating a customer location (for rental or sale usage), it
is now returned with a sudo environment. This is a bad thing as
its environment is reused to create and validate the picking in
the send_devices contract method, so it may be used to bypass
the necessary permissions for this operation.