[test] Assigned desirable groups for first B2B user with portal granting
Since user1 is the first user in the objects' company to be granted portal access,
they also get granted all of the customer roles (in the customer_team_manager module),
including customer admin, invoice and tasks roles.
As such, they don't fulfil the requirement of a random user which doesn't have these roles.

To override this behavior, we directly assign the required role to the user after being assigned.
Afterwards, since the company already has users with a granted portal access, they don't obtain
the roles.