Added date_range_portal to dependencies
Since the customer manager modules give a limited access to the backend for portal users,
some records being loaded may not be accessible, even to read, for portal user, understandably so.

Some of these models is date.range and date.range.type, accessible through the customer_device_manager view,
which can trigger an access error for a portal user.
As such, we add this newly created module as a dependency of customer_device_manager.