Fixed modification allowed field for partner image
To follow the implementation of partner avatars in the Contacts module, we displayed the image_1920 field
using a lower resolution (avatar_128).

However, within that migration, we modified the wrong image field in the fields which can be modified by customer admins,
which caused an AccessError with the _check_customer_allowed_attrs method.
As such, we change the modified field in order to properly allow admins to modify their employees' data without raising an Access Error.