[test] Revamped partner creation by customer admin using Form
While creating a partner directly through the model create method with the correct fields works well,
it might fail when creating a partner through the form view.
Indeed, as the form view passes other fields through the create method, like a default profile picture (image_1920)
or a company status (is_company), these fields might make the check fail, even though they might be accepted (depending on valid values)

As such, to directly test a form creation of a partner as a customer admin, we use the Form class which allows to check
for potential unchecked fields, and will make this feature more robust.