Added Repackaged devices as a valid location to select stock quants
At the moment, it's only possible to select devices which are either in the 'New' or 'To check' locations.
However, to prepare a device which was received already Refurbished, this can cause an issue to initially
select the device's lot.

As such, we expand the range of selectable quants to fetch more devices in the task lot_id field