Added duplicate out-of-stock notif. verification for picking wizard
During the products locations computation in a wizard, if any product is out-of-stock,
we send a notification to the user in order to inform them of the stock issue.

However, the products_locations field is computed twice upon creation, since it is linked
to the onchange_all_products_or_priority method, and upon creation, both all_products and
prioritize_repackaged are written.

As such, when sending a notification for a out-of-stock product, we now check if other notifications
were sent with the same message or not, and unsend it in that case.