[IMP] Replaced server.env.mixin by server.env.techname.mixin
Previously, we used server.env.mixin to apply env. variables through the configuration file,
which fetched the data from their name attribute.

However, the name attribute can't possess spaces, which can more easily be changed,
and lead to readability issues from the front-end.

As such, we chose to inherit from server.env.techname.mixin, which allows the assignment of a tech_name field,
which is more reliable overall and can be more complicated.