[IMP] Changed XPath expr. in portal menu entry inheritences to use URLs
Beforehand, we based the modification of the portal menu entry titles on the titles themselves.
However, since these titles can be translated, and we based the XPath expressions on their original (english) name,
then switching to another language (for instance: french) would make the XPath expressions fail.

As such, we switched to using the 'url' values, which are static in any languages, and as such no longer
prone to this type of issue.