[FIX] Rearranged XPath attribute element
- The first XPath insertion aimed at a single element in the div element containing
the product price, in the products list.

```html
<!-- website_sale/views/templates.xml - L193-->
<div itemprop='offers'>
  <t>
    <!-- ... -->
  </t>
  <span .../>
  <span .../>
</div>
```

Since we wished to replace all of the tags contained within the div element,
and not only the t element, we now insert our new element after the div element,
and we wish to remove the whole div element, rather than a single <b> element,
like how it was in v12.

- The second replacement was aiming at the <b> element contained within the div element.
Since we added previously that <b> element, this XPath expression didn't crash, however
since it didn't contain a 'position' attribute and didn't contain elements within it,
this had no effect.

As such, we aim directly at the parent div, which contains all of the elements
we wish to remove from the view, and set it to never appear.