[mig] Migration invoice_id column of project tasks
In 12.0, this column used to point to a account.invoice instances, which
have been replaced by account.move. In this commit, we update each project
task pointing to an invoice so that the invoice_id column is related to the
account.move which replaces the old account.invoice instance, that is to say
the one matching the old_invoice_id field.