Restrict selectable partner type for customer device assignment to 'contact' and 'other'
This makes it more coherent with the customer_team_manager partner list.
Note that a normal difference is that a company can be assigned to a
device, which signifies the device is not assigned inside the company but
may be used as a spare device in case of an emergency for instance.