[FIX] Fixed XPath element attribute modification
The goal of this XPath expression was to hide an element in the targeted template.
However, since the XPath directive didn't contain any instructions or elements within,
and the 't-if' attribute was on the XPath element itself, it didn't do anything,
although the element was correctly selected (ie. no crash at QWeb validation).