[IMP] commown_self_troubleshooting: Changed target project for Other-type issues
In commown_contractual_issue's project.task view, the contract_id field is required if the task's project is configured
to mandate contracts, like the Support project.
However, the Other issue self-troubleshooting track created a task without contract in this project.

Editing a Other-type issue in the Support project is usually impossible without a contract assigned.
To bypass this requirement for customers without contracts, the Support team used a glitch present in Odoo 12.0,
which allowed the editing of a task's stage using the statusbar element at the top of a task.

However, with Odoo 16.0's constant Edit mode activated, editing all fields, including changing the task's stage
using the statusbar is impossible.

As such, the internal decision made was to create newly Other-type issues in the Commercial Support project,
which doesn't require a contract on any tasks.