[FIX] commown_self_troubleshooting: (tests) Fixed DOM element selection for customer name
Between Odoo 12.0 and Odoo 16.0, in the customer portal DOM, a bold HTML element wrapping the customer name was removed.
This provoked an issue in the name text fetch to check the whether the pre-inputed data in the self-troubleshoot form
matched the logged in user.

So, we remove this bold element to be able to fetch the name value in the DOM.