[IMP] commown_shipping: Added dedicated group to access print label wizards
Since an Odoo warning is sent when a model has no set access rules, we assigned a group from the product module
to be able to access the wizard models.

However, since the concerned group (product.group_stock_packaging) also affects the display of other fields
related to the product.packaging record, which we don't use, we decided to create a dedicated group
to access these printing label wizards.