[MIG] commown_devices: Renamed origin field conflicting with new stock.move.line field
In Odoo v16, the origin field was added in the stock.move.line model, which conflicts
with our added field of the same name.
This conflicts with our compute method, since the Odoo field has a related attribute overriding it.
As such, we renamed our added field.

Additionnally, we modify the related test, since the compute method should work upon field access,
and not only when the compute method is called, which didn't allow us to detect this glitch.