[IMP] commown_devices: Removed group elements around tree fields
Within a group element, children nodes are rendered in a grid, with at minimum 2 columns.
This means that nested tree views are arranged in a singular column by default,
which can make these tree views, and their columns narrow.

As such, we removed the group attributes, and display the nested tree views one after the other,
while adding a separator element to label each view.