[IMP] Added contract view inheritance to replace button strings by titles
In the button rendering templates, 'string' values are used to display a label next to any given button,
while the 'title' value is used to pass as a button's title (displayed when hovering the cursor on top of the button).

Since the button labels take a lot of space, users prefer to insert a button's function in the title rather than the string.