[FIX] commown_devices: Added date_done assignment in _force_scrap_date
The related action's purpose is to set both a scrap's date_done field and related move/quants' date fields
to the move line date (see commit 9b53a94b6 )

However, the date_done field was not assigned in the _force_scrap_date function, which lead to updated moves/quants,
but the scrap itself showing the same date as the validation date.

We also modified the related test, with an updated comment and a new assert on the scrap.date_done field.