[FIX] website_sale_b2b: Fixed company_name addition to field_required field
In the commit 169bc7390, this XPath element was extracted from the commown module into the website_sale_b2b.
During the migration, the string evaluation was replaced by a t-if condition which weren't taken into account,
as the XPath expression always added company_name to the field_required field, even if edit_company_infos was wrong.

As such, we revert back the usage of edit_company_infos boolean in the string evaluation.