[IMP] contract_variable_discount: Added name_get method to compute display_name
In the v12 implementation of the product_contract module, a display_name compute method was added in the contract.line model.
However, this method is gone in the v16 version of the module.

As such, we implemented a name_get method which computes the same display_name as product_contract's v12 display_name compute method.