[FIX] commown_automated_control: Added automated_control_id assignment if automated_control_ids is null
Since Odoo expects the automated_control_id value to be assigned to a base.automation record through the compute method,
if the automated_control_ids field is null, the field is currently not assigned, leading to a CacheMiss exception.

As such, we assign a False value to the field, if the given automation record's automated_control_ids field is null.
We also add a test to certify the code's functionnality.