[IMP] account_invoice_merge_auto_pay: Added default False value to auto_merge for newly created reserve moves
To create a reversed move of a given move, the reverse_moves method copies the original move, to get the same fields,
except for values set in the default_values_list argument.

However, in the _reverse_moves method overload of the account_payment_partner, a payment_mode_id is set, to try to fetch a refund payment mode
(move.payment_mode_id.refund_payment_mode_id), which can be False.

This causes an issue when the record is created, since if the auto_merge value of the original move is set to True, and the payment_mode_id value is False,
the _check_auto_merge method raises an error.

As such, for newly created reverse moves, we set the auto_merge value to False.