Fix wrong stage for new sale followup tasks
The followup tasks created when the sale of a contract product with a
'followup_sales_project_id' non-empy field is confirmed, was created
with an undefined stage_id. The method to choose this stage was not
adapted to the project.task stage case but only to the crm.lead case.

This is now fixed, and a test was added.