[IMP] website_sale_b2b: Changed the default add_qty value to be an integer
Since we sell devices and accessories, it's not possible to order half of a device (three-and-a-half Fairphones, for instance).
However, on the product page, it's possible to set a floating value, and the default value is a float, meaning the first decimal is shown (ie. "10.0")

While the quantity gets rounded to an integer when added to the cart, it can still mislead users on the store.
As such, we force the default add_qty value to an integer when passing the default add_qty value to the product template.