[FIX] commown_self_troubleshooting: Added explicit non-user assignment of newly created self-troubleshoot tasks
In Odoo v12, when creating a project.task through the create_method, the default values of the model (default_get)
didn't assign a default value to user_id.

However, in Odoo v16, the same method now assigns the current environnement user as the default assignee to the newly created task,
who can be a customer when a task is created through the self-troubleshooting forms.

As such, we explicitly set the user_ids value to False, to avoid having a customer be assigned to a newly created task.