[IMP] commown_contractual_issue: Added default_contract_type context value on project.task contract_id field
When browsing a contract and clicking on the Invoices action, the action computes
the default_move_type context value based on the default_contract_type value :

- 'sale' -> 'out_invoice'
- Otherwise -> 'in_invoice'
(see contract/contract/models/contract.py:L282)

The default_move_type context value is then used to tweak action behavior, as well as the domain
of certain fields within the view - for instance, which products are available to assign on a move line.

The default_contract_type value is set when browsing from Accounting -> Customers -> Customer contracts,
but not when going on a contract from a project.task - as such, we add this context value when
accessing a contract from a project.task.